Do not loop indefinitely when doing getAbi.
Previously a call to getAbi where the node.h and node_version.h
files cannot be found (due to a download failure or logic error)
would cause prebuild to loop indefinitely trying to download that
release.  This causes prebuild to only attempt to download the
files a single time before failing.